Building onto the flask website

+++ host/flask/classlib/database.py
@@ -1,7 +1,10 @@
+from collections import Counter
+from datetime import datetime
+
 import mysql.connector
 
 
-class Database():
+class Database:
     def __init__(self):
         self.connection = mysql.connector.connect(
             host="35.201.1.71",
@@ -11,7 +14,8 @@ class Database():
 
     def getAllBooks(self):
         cursor = self.connection.cursor()
-        cursor.execute("SELECT * FROM book;")
+        cursor.execute(
+            "SELECT book.*, bookBorrowed.status FROM book LEFT JOIN bookBorrowed ON bookBorrowed.bookID = book.bookID;")
         return cursor.fetchall()
 
     def addBook(self, title, author, publishedDate):
@@ -22,11 +26,44 @@ class Database():
 
     def deleteBook(self, bookID):
         cursor = self.connection.cursor()
-        cursor.execute("DELETE FROM bookBorrowed WHERE bookID=%s", (bookID,))
-        cursor.execute("DELETE FROM book WHERE bookID=%s", (bookID,))
-        self.connection.commit()
+        cursor.execute("SELECT status FROM bookBorrowed WHERE bookID=%s ORDER BY borrowDate DESC;", (bookID,))
+        row = cursor.fetchone()
+        borrowed = row[0]
+        if borrowed == "BORROWED":
+            return False
+        if borrowed == "RETURNED":
+            cursor.execute("DELETE FROM bookBorrowed WHERE bookID=%s;", (bookID,))
+            cursor.execute("DELETE FROM book WHERE bookID=%s;", (bookID,))
+            self.connection.commit()
+            return True
 
     def getOneBook(self, bookID):
         cursor = self.connection.cursor()
-        cursor.execute("SELECT * FROM book WHERE bookID=%s;", (bookID,))
+        cursor.execute(
+            "SELECT book.*, COUNT(bookBorrowed.bookID) as borrowedAmount FROM book LEFT JOIN bookBorrowed ON bookBorrowed.bookID = book.bookID WHERE book.bookID=%s GROUP BY book.bookID;",
+            (bookID,))
         return cursor.fetchone()
+
+    def getBookAverageBorrowData(self, bookID):
+        cursor = self.connection.cursor()
+        cursor.execute("SELECT borrowDate FROM bookBorrowed WHERE bookID=%s", (bookID,))
+        rows = cursor.fetchall()
+        counter = Counter()
+
+        for row in rows:
+            if datetime.strptime(rows[0], "%Y-%M-%D").weekday() == 0:
+                counter["monday"] += 1
+            elif datetime.strptime(rows[0], "%Y-%M-%D").weekday() == 1:
+                counter["tuesday"] += 1
+            elif datetime.strptime(rows[0], "%Y-%M-%D").weekday() == 2:
+                counter["wednesday"] += 1
+            elif datetime.strptime(rows[0], "%Y-%M-%D").weekday() == 3:
+                counter["thursday"] += 1
+            elif datetime.strptime(rows[0], "%Y-%M-%D").weekday() == 4:
+                counter["friday"] += 1
+            elif datetime.strptime(rows[0], "%Y-%M-%D").weekday() == 5:
+                counter["saturday"] += 1
+            elif datetime.strptime(rows[0], "%Y-%M-%D").weekday() == 6:
+                counter["sunday"] += 1
+
+        return counter
